Wiki: Correct PixPilot serial connection for Crossfire
The page currently recommends using SERIAL4 = UART4 (aka GPS2) for Crossfire and F-Port, but if you do that ArduPilot throws an warning: "CRSFv2" running on non-DMA serial port", because UART4 only has DMA on Transmit (it is not a "full DMA port") - the only port with bi-directional DMA is SERIAL2 = UART2. I have tested this with 4.4.0 and it works on SERIAL2.